requirements.txt
asgiref==3.9.1
astroid==3.3.11
attrs==25.3.0
brotli==1.2.0
certifi==2025.8.3
colorama==0.4.6
coverage==7.10.6
dill==0.4.0
Django==5.2.4
django-cors-headers==4.9.0
django-debug-toolbar==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
docopt==0.6.2
dotenv==0.9.9
env==0.1.0
exceptiongroup==1.3.0
Faker==37.4.2
gunicorn==23.0.0
h11==0.16.0
idna==3.10
iniconfig==2.1.0
isort==6.0.1
mccabe==0.7.0
outcome==1.3.0.post0
packaging==25.0
parameterized==0.9.0
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
psycopg==3.2.12
psycopg2-binary==2.9.11
Pygments==2.19.2
PyJWT==2.10.1
pylint==3.3.7
PySocks==1.7.1
pytest==8.4.1
pytest-django==4.11.1
pytest-dotenv==0.5.2
pytest-watch==4.2.0
python-dotenv==1.1.1
selenium==4.35.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
tomli==2.2.1
tomlkit==0.13.3
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
websocket-client==1.8.0
wsproto==1.2.0
